feat(custom-items): reimplement all custom items
