+++ src/main/java/me/lutto/shardsmp/items/weapons/PoseidonTrident.kt
@@ -1,15 +1,30 @@
-package me.lutto.shardsmp.listeners.weapons
+package me.lutto.shardsmp.items.weapons
 
 import me.lutto.shardsmp.items.events.AbilityActivateEvent
 import me.lutto.shardsmp.ShardSMP
-import org.bukkit.Location
-import org.bukkit.Material
-import org.bukkit.World
+import me.lutto.shardsmp.items.CustomCooldownItem
+import net.kyori.adventure.text.format.TextDecoration
+import net.kyori.adventure.text.minimessage.MiniMessage
+import org.bukkit.*
 import org.bukkit.block.data.Waterlogged
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 
-class PoseidonTridentListener(private val shardSMP: ShardSMP) : Listener {
+class PoseidonTrident(private val shardSMP: ShardSMP) : CustomCooldownItem(
+    "poseidon_trident",
+    Material.TRIDENT,
+    MiniMessage.miniMessage().deserialize("<gradient:#1616aa:#3646ff>Poseidon's Trident")
+        .decoration(TextDecoration.ITALIC, false),
+    listOf(MiniMessage.miniMessage().deserialize("<gold>[Shift + Left Click]").decoration(TextDecoration.ITALIC, false)),
+    11,
+    false,
+    180,
+    true
+), Listener {
+
+    init {
+        shardSMP.itemManager.registerItem(this)
+    }
 
     @EventHandler
     fun onAbilityActivate(event: AbilityActivateEvent) {
